requirements.txt
langchain==0.3.0
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.38.0
//openai==1.44.1
faiss-cpu==1.8.0.post1
huggingface-hub==0.24.6
altair==5.4.1
InstructorEmbedding==1.0.1
sentence-transformers==3.1.0


